+++ CHANGELOG.md
@@ -4,5 +4,8 @@
 ### Added
 - Added colour palette to plugin settings.
 
+### Changed
+- General UI improvements.
+
 ## 1.0.0 - 2020-03-09
 - Initial release.