+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes for Plugin Sales
  
-## 1.1.0 - Unreleased
+## 1.1.0 - 2020-03-10
 ### Added
 - Added a colour palette to plugin settings.
 - Added console commands to refresh and delete sales.